extracted centralities for mln into another script
