+++ docs/source/reference_guide.rst
@@ -6,6 +6,7 @@ Reference guide
    :maxdepth: 3
 
    network_models
+   net_collection
    spreading_models
    seed_selection
    performing_experiments
