+++ network_diffusion/nets/toy_network_piotr.py
@@ -63,6 +63,3 @@ def get_toy_network_piotr() -> MultilayerNetwork:
     return MultilayerNetwork.from_nx_layers(
         [layer_1, layer_2, layer_3], ["l1", "l2", "l3"]
     )
-
-
-# TODO: add sphinx docs
