+++ network_diffusion/seeding/degreecentrality_selector.py
@@ -13,11 +13,11 @@ from typing import Any
 import networkx as nx
 
 from network_diffusion.mln.actor import MLNetworkActor
+from network_diffusion.mln.centralities import degree
 from network_diffusion.mln.centrality_discount import (
     degree_discount,
     degree_discount_networkx,
 )
-from network_diffusion.mln.functions import degree
 from network_diffusion.mln.mlnetwork import MultilayerNetwork
 from network_diffusion.seeding.base_selector import BaseSeedSelector
 from network_diffusion.utils import BOLD_UNDERLINE, THIN_UNDERLINE
