+++ network_diffusion/seeding/kshell_selector.py
@@ -13,7 +13,7 @@ from typing import Any
 import networkx as nx
 
 from network_diffusion.mln.actor import MLNetworkActor
-from network_diffusion.mln.functions import degree, k_shell_mln
+from network_diffusion.mln.centralities import degree, k_shell_mln
 from network_diffusion.mln.mlnetwork import MultilayerNetwork
 from network_diffusion.seeding.base_selector import (
     BaseSeedSelector,
