+++ network_diffusion/seeding/neighbourhoodsize_selector.py
@@ -13,10 +13,10 @@ from typing import Any
 import networkx as nx
 
 from network_diffusion.mln.actor import MLNetworkActor
+from network_diffusion.mln.centralities import neighbourhood_size
 from network_diffusion.mln.centrality_discount import (
     neighbourhood_size_discount,
 )
-from network_diffusion.mln.functions import neighbourhood_size
 from network_diffusion.mln.mlnetwork import MultilayerNetwork
 from network_diffusion.seeding.base_selector import BaseSeedSelector
 from network_diffusion.utils import BOLD_UNDERLINE, THIN_UNDERLINE
