+++ network_diffusion/tests/models/test_mlt_model.py
@@ -1,7 +1,7 @@
 import unittest
 
-from network_diffusion.mln.functions import get_toy_network_piotr
 from network_diffusion.models import MLTModel
+from network_diffusion.nets import get_toy_network_piotr
 from network_diffusion.seeding.mocking_selector import MockingActorSelector
 from network_diffusion.simulator import Simulator
 from network_diffusion.utils import fix_random_seed
